requirements.txt
wheel
Django
django-environ
djangorestframework
djangorestframework-simplejwt
mysqlclient
autopep8
uuid
django-cors-headers